+++ changelog.md
@@ -2,12 +2,13 @@
 
 - New option to display the original image (without color adaptation) as a preview;
 - New option for filling transparent background with a specific map color;
+- The original image now can be scaled to a size smaller than the mapart;
 
 <details>
 <summary>Fixes and minor changes</summary>
 
 - Staircase algorithms has been rewritten and optimized;
-- Added displaying of the color names in the preset editor; 
+- Added displaying of the color names in the preset editor;
 - Fixed reading air and unexisting blocks from complete_palette.json;
 - Fixed mouse scroll behavior in the preset editor above and below the window box;
 - Added blocking of some buttons which requires mapart with color adaptation;